chore: separated all translations models to own files, removed translations input type
