+++ src/models/string-or-translations.ts
@@ -1,0 +1,14 @@
+import Joi from 'joi';
+import { schema as translationsSchema, Translations } from './translations';
+
+const schema = Joi.alternatives(
+  Joi.string(),
+  translationsSchema,
+);
+
+type StringOrTranslations = string | Translations;
+
+export {
+  schema,
+  StringOrTranslations,
+};
