+++ src/models/translations.ts
@@ -1,24 +1,16 @@
 import Joi from 'joi';
 
-const translationsSchema = Joi.object().keys({
+const schema = Joi.object().keys({
   en: Joi.string().required(),
   nl: Joi.string().required(),
 });
-const schema = Joi.alternatives(
-  Joi.string(),
-  translationsSchema,
-);
 
 interface Translations {
   en: string,
   nl: string,
 }
 
-type StringOrTranslations = string | Translations;
-
 export {
   schema,
-  translationsSchema,
   Translations,
-  StringOrTranslations,
 };
